math-solver-full (9)/math-solver-full (7)/math-solver-full/requirements.txt
pandas==2.0.3
numpy==1.24.3
sentence-transformers==2.2.2
scikit-learn==1.3.0
openai==1.3.0
anthropic==0.7.0
pyyaml==6.0.1
torch==2.0.1
transformers==4.35.0
faiss-cpu==1.7.4
regex==2023.10.3
tqdm==4.66.1
